% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headcount.R
\name{get_headcount}
\alias{get_headcount}
\title{Get Headcount}
\usage{
get_headcount(.data, date_val)
}
\arguments{
\item{.data}{Existing data. Not required if piping (\verb{\%>\%}) using the tidyverse workflow.}

\item{date_val}{Date for the time that headcount should be calculated. Date should be in ymd format. It is recommended that all date field are correctly declared using lubridate::ymd().}
}
\value{
A numeric vector.
}
\description{
Get Headcount
}
