% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headcount_avg.R
\name{get_headcount_avg}
\alias{get_headcount_avg}
\title{Get Headcount Average}
\usage{
get_headcount_avg(.data, d1, month_offset = 3)
}
\arguments{
\item{.data}{Existing data. Not required if piping (\verb{\%>\%}) using the tidyverse workflow.}

\item{d1}{Date for the time that headcount should be calculated. Date should be in ymd format. It is recommended that all date field are correctly declared using lubridate::ymd().}

\item{month_offset}{Number of \emph{months} to offset the original date. This is set to 3 months by default (the span of 1 quarter.)}
}
\value{
Tibble containing headcounts at two dates plus the average headcount.
}
\description{
Get Headcount Average
}
\examples{
# Standard usage.
wfm_tms_play \%>\% 
  get_headcount_avg("2022-01-01")
  
# Using the group_by() dplyr function.  
wfm_tms_play \%>\% 
  dplyr::group_by(region) \%>\% 
  get_headcount_avg("2022-01-01")
}
