% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headcount.R
\name{get_headcount}
\alias{get_headcount}
\title{Get Headcount}
\usage{
get_headcount(.data, date_val)
}
\arguments{
\item{.data}{Existing data. Not required if piping (\verb{\%>\%}) using the tidyverse workflow.}

\item{date_val}{Date for the time that headcount should be calculated. Date should be in ymd format. It is recommended that all date field are correctly declared using lubridate::ymd().}
}
\value{
A single numeric value in a 1x1 tibble matrix or a tibble matrix that takes into account any \code{group_by()} actions.
}
\description{
Get Headcount
}
\examples{
# Standard usage.
wfm_tms_play \%>\% 
  get_headcount("2019-01-01")
  
# Using the group_by() dplyr function.  
wfm_tms_play \%>\% 
  dplyr::group_by(region) \%>\% 
  get_headcount("2019-01-01")
}
