% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wfm_tms_play}
\alias{wfm_tms_play}
\title{Dummy dataset mimicking all_tms}
\format{
A data frame with 50000 rows and 11 variables:
\describe{
\item{tm_name}{Fictitious employee ID.}
\item{region}{Fictitious region of employment.}
\item{team}{Fictitious region of employment.}
\item{hire_date}{Fictitious region of employment.}
\item{termination_date}{Fictitious region of employment.}
\item{termination_category}{Fictitious region of employment.}
\item{termination_reason}{Fictitious region of employment.}
\item{tenure}{Fictitious region of employment.}
\item{gender}{Fictitious region of employment.}
\item{is_employee_active}{Fictitious region of employment.}
\item{most_recent_job_satisfaction_score}{Fictitious region of employment.}
}
}
\usage{
wfm_tms_play
}
\description{
A dataset hire date, termination date, and the like. Fabricated.
}
\keyword{datasets}
